fix: bug in auth controller(gen refresh token)

+++ backend/sswm/src/main/java/com/ground/sswm/auth/domain/AuthRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthRepository extends JpaRepository<Auth,Integer> {
+public interface AuthRepository extends JpaRepository<Auth,Long> {
     Auth findByUserId(Long userId);
 }
